+++ src/modules/app/index.js
@@ -1,6 +1,5 @@
 /** @module modules/app */
 export * from './ElementModule';
-export * from './CameraModule';
 export * from './RenderingModule';
 export * from './SceneModule';
 export * from './ResizeModule';
@@ -9,3 +8,6 @@ export * from './VirtualMouseModule';
 export * from './EventsPatchModule';
 export * from './ControlsModule';
 export * from './FogModule';
+
+// controls
+export * from './controls/index';
