+++ src/modules/index.js
@@ -1,0 +1,4 @@
+/** @module modules */
+
+export * from './app/index';
+export * from './mesh/index';
