--- src/modules/presets/index.js
@@ -1,1 +1,0 @@
-export * from './app/basic';
